Make the new downloads page the default.
