+++ source/assets/javascripts/all.js
@@ -7,7 +7,6 @@
 //= require 'bootstrap-sprockets'
 
 //= require 'app/_main'
-//= require 'app/_download_new'
 //= require 'app/_download'
 //= require 'app/_getting_started'
 //= require 'app/_tabs'
