+++ source/assets/stylesheets/site.css.scss
@@ -29,7 +29,6 @@
 @import "scss/help";
 @import "scss/getting-started";
 @import "scss/blog";
-@import "scss/downloads_new";
 @import "scss/downloads";
 @import "scss/events";
 @import "scss/why-gocd";
